+++ Sources/SheetMusicLayout/Layout/LayoutEngine+YBounds.swift
@@ -67,8 +67,8 @@ extension LayoutEngine {
              let .multiMeasureRest(_, p),
              let .measureNumber(_, p),
              let .staffName(_, p),
-             let .staffText(_, p, _, _),
-             let .rehearsalMark(_, p, _, _),
+             let .staffText(_, p, _, _, _),
+             let .rehearsalMark(_, p, _, _, _),
              let .rest(_, p, _, _, _),
              let .note(_, _, _, _, p, _, _, _):
             return [p.y]